pointsGenerator moved to package
