+++ TagCloud/RectanglesLayouter/PointsGenerator/IPointsGenerator.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Drawing;
 
-namespace TagCloud
+namespace TagCloud.RectanglesLayouter.PointsGenerator
 {
     public interface IPointsGenerator
     {
