+++ TagCloud/RectanglesLayouter/PointsGenerator/SpiralPointsGenerator.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Drawing;
 
-namespace TagCloud
+namespace TagCloud.RectanglesLayouter.PointsGenerator
 {
     public class SpiralPointsGenerator : IPointsGenerator
     {
